Convert GemTrader plugin from kts to kt
