+++ game/plugins/src/main/kotlin/gg/rsmod/plugins/content/areas/alkharid/GemTrader.kt
@@ -1,0 +1,53 @@
+package gg.rsmod.plugins.content.areas.alkharid
+
+import gg.rsmod.game.Server
+import gg.rsmod.game.model.World
+import gg.rsmod.game.model.queue.QueueTask
+import gg.rsmod.game.model.shop.PurchasePolicy
+import gg.rsmod.game.model.shop.ShopItem
+import gg.rsmod.game.plugin.KotlinPlugin
+import gg.rsmod.game.plugin.PluginRepository
+import gg.rsmod.plugins.api.cfg.Items
+import gg.rsmod.plugins.api.cfg.Npcs
+import gg.rsmod.plugins.api.ext.*
+import gg.rsmod.plugins.content.mechanics.shops.CoinCurrency
+
+class GemTrader(r: PluginRepository, world: World, server: Server) : KotlinPlugin(r, world, server) {
+
+    init {
+        create_shop(
+            "Gem Trader",
+            currency = CoinCurrency(),
+            purchasePolicy = PurchasePolicy.BUY_STOCK,
+            containsSamples = false,
+        ) {
+            items[0] = ShopItem(Items.UNCUT_SAPPHIRE, 0)
+            items[1] = ShopItem(Items.UNCUT_EMERALD, 0)
+            items[2] = ShopItem(Items.UNCUT_RUBY, 0)
+            items[3] = ShopItem(Items.UNCUT_DIAMOND, 0)
+            items[4] = ShopItem(Items.SAPPHIRE, 0)
+            items[5] = ShopItem(Items.EMERALD, 0)
+            items[6] = ShopItem(Items.RUBY, 0)
+            items[7] = ShopItem(Items.DIAMOND, 0)
+        }
+
+        on_npc_option(Npcs.GEM_TRADER, "trade") {
+            player.openShop("Gem Trader")
+        }
+
+        on_npc_option(Npcs.GEM_TRADER, "talk-to") {
+            player.queue { chat(this) }
+        }
+    }
+
+    suspend fun chat(it: QueueTask) {
+        it.chatNpc("Good day to you traveller.", "Would you be interested in buying some gems?")
+        when (it.options("Yes please.", "No thank you.")) {
+            1 -> it.player.openShop("Gem Trader")
+            2 -> {
+                it.chatPlayer("No thank you.")
+                it.chatNpc("Eh, suit yourself.")
+            }
+        }
+    }
+}
